Refactor LabelPositions struct into a Position enum.
The Position enum uses an associated value, of type PositionTuple
(which is a typealias for  (xMultiplier: Float, yMultiplier: Float)).
In order to get easy access to the associated value of the case the
positionTuple property is provided. Finally the staticX (Center, Right,
etc) is provided for static access to the filled in associated values,
specific to TickerView labels.