Make SegmentType (primarily) a value type.
While types like CountDownSegment, OvertimeSegment, etc, were structs
they returned a Timer<T> class, this created a fake value type. Fake
value types presented an advantage and a disadvantage. The advantage is
that fake value types allowed for the remarkably easy modeling of a
global timer appearing multiple times. Unfortunately it brought with it
the costs of a reference type; primarily it’s impossible for a round,
once completed to reset to it’s original state, a behavior that’s
needed, since at the end of the round the most logical way to finish is
to start over.